Perform stringification in the code generator before other #includes.
We don't want macros defined in the standard headers to apply to our
stringified files.  For example, memset was becoming the GNU-ism
"__builtin___memset_chk" on some platforms, which isn't appropriate for a
cross-platform header.

+++ x-construct-actions.h
@@ -28,7 +28,7 @@
 
 #define CONSTRUCT_IS_END_ACTION(action) (CONSTRUCT_ACTION_GET_TYPE(action) & 8)
 
-static inline const char *action_name(uint16_t action)
+static inline const char *action_name(unsigned action)
 {
     switch (CONSTRUCT_ACTION_GET_TYPE(action)) {
     case ACTION_NONE:
